+++ examples/basic-tag/app/index.js
@@ -1,0 +1,4 @@
+var oval = require('organic-oval')
+oval.init()
+require('./navigation')
+oval.mountAll('*', document)